+++ TR5Main/Objects/TR2/Entity/tr2_monk.cpp
@@ -35,9 +35,9 @@ void MonkControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short torso = 0;
-	short angle = 0;
-	short tilt = 0;
+	float torso = 0;
+	float angle = 0;
+	float tilt = 0;
 
 	if (item->HitPoints <= 0)
 	{
