+++ TR5Main/Objects/TR2/Entity/tr2_rat.cpp
@@ -33,9 +33,9 @@ void RatControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* info = GetCreatureInfo(item);
 
-	short head = 0;
-	short angle = 0;
-	short random = 0;
+	float head = 0;
+	float angle = 0;
+	float random = 0;
 
 	if (item->HitPoints <= 0)
 	{
