+++ TR5Main/Objects/TR3/Entity/tr3_cobra.cpp
@@ -44,9 +44,9 @@ void CobraControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* info = GetCreatureInfo(item);
 
-	short head = 0;
-	short angle = 0;
-	short tilt = 0;
+	float head = 0;
+	float angle = 0;
+	float tilt = 0;
 
 	if (item->HitPoints <= 0)
 	{
