+++ TR5Main/Objects/TR4/Entity/tr4_harpy.cpp
@@ -250,14 +250,14 @@ void HarpyControl(short itemNumber)
 	auto* creature = GetCreatureInfo(item);
 	auto* object = &Objects[item->ObjectNumber];
 
-	short angle = 0;
-	short joint0 = 0;
-	short joint1 = 0;
-	short joint2 = 0;
+	float angle = 0;
+	float joint0 = 0;
+	float joint1 = 0;
+	float joint2 = 0;
 
 	if (item->HitPoints <= 0)
 	{
-		short state = item->Animation.ActiveState - 9;
+		int state = item->Animation.ActiveState - 9;
 		item->HitPoints = 0;
 
 		if (state)
