+++ TR5Main/Objects/TR4/Entity/tr4_mummy.cpp
@@ -84,11 +84,11 @@ void MummyControl(short itemNumber)
 	auto* item = &g_Level.Items[itemNumber];
 	auto* creature = GetCreatureInfo(item);
 
-	short tilt = 0;
-	short angle = 0;
-	short joint0 = 0;
-	short joint1 = 0;
-	short joint2 = 0;
+	float tilt = 0;
+	float angle = 0;
+	float joint0 = 0;
+	float joint1 = 0;
+	float joint2 = 0;
 
 	if (item->AIBits)
 		GetAITarget(creature);
