+++ TR5Main/Objects/TR4/Entity/tr4_setha.cpp
@@ -76,7 +76,7 @@ void SethaControl(short itemNumber)
 	int height4 = GetFloorHeight(floor, x, y, z);
 
 	AI_INFO info;
-	short angle = 0;
+	float angle = 0;
 
 	if (item->HitPoints <= 0)
 	{
